staging/ft1000-usb: fix unlocked_ioctl prototype
unlocked_ioctl has a "long" return type.

+++ drivers/staging/ft1000/ft1000-usb/ft1000_chdev.c
@@ -51,7 +51,7 @@ extern void CardSendCommand(struct ft1000_device *ft1000dev, unsigned short *pte
 
 static int ft1000_ChOpen (struct inode *Inode, struct file *File);
 static unsigned int ft1000_ChPoll(struct file *file, poll_table *wait);
-static int ft1000_ChIoctl(struct file *File, unsigned int Command,
+static long ft1000_ChIoctl(struct file *File, unsigned int Command,
                            unsigned long Argument);
 static int ft1000_ChRelease (struct inode *Inode, struct file *File);
 
@@ -536,7 +536,7 @@ static unsigned int ft1000_ChPoll(struct file *file, poll_table *wait)
 // Notes:
 //
 //---------------------------------------------------------------------------
-static int ft1000_ChIoctl (struct file *File, unsigned int Command,
+static long ft1000_ChIoctl (struct file *File, unsigned int Command,
                            unsigned long Argument)
 {
     struct net_device *dev;
